Support for 64-bit Platform
* All: [BUG FIX] Corrected the tests that are failed by the pointer sizes and alignments are changed.
* All: [BUG FIX] Corrected 'warning C4244'.
* All: The class StackTrace will become that makes the stack trace with using RtlCaptureContext api, because the inline assembler is not supported in the 64-bit platform.
* Collections: Make the performance of RapidVector to same as the 32-bit.

+++ Test.Urasandesu.CppAnonym/Test/Urasandesu/CppAnonym/Utilities/VariantPtrTest.cpp
@@ -75,7 +75,7 @@ namespace {
 
         typedef VariantPtr<Hoge *, intrusive_ptr<Hoge> > MyVariantPtr;
 #ifdef _DEBUG
-        BOOST_MPL_ASSERT_RELATION(sizeof(MyVariantPtr), ==, sizeof(intrusive_ptr<Hoge>) + sizeof(INT));
+        BOOST_MPL_ASSERT_RELATION(sizeof(MyVariantPtr), ==, sizeof(intrusive_ptr<Hoge>) + sizeof(SIZE_T));
 #else
         BOOST_MPL_ASSERT_RELATION(sizeof(MyVariantPtr), ==, sizeof(intrusive_ptr<Hoge>));
 #endif
